requirements.txt
click==8.1.7
python-gitlab==4.4.0